requirements.txt
requests>=2.31.0
spacy>=3.7.0
geopy>=2.4.1
pyyaml>=6.0.1
python-dateutil>=2.9.0
tenacity>=8.2.3
pytest>=8.1.0
